requirements.txt
pypdf2
azure-storage-blob
mongoengine
pymongo
python-dotenv
cairosvg
magic